% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{BlindWriteRequest}
\alias{BlindWriteRequest}
\title{BlindWriteRequest Object}
\usage{
BlindWriteRequest(mutation = NULL)
}
\arguments{
\item{mutation}{The mutation to perform}
}
\value{
BlindWriteRequest object
}
\description{
BlindWriteRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other BlindWriteRequest functions: \code{\link{datasets.blindWrite}}
}
