% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{BlindWriteResponse}
\alias{BlindWriteResponse}
\title{BlindWriteResponse Object}
\usage{
BlindWriteResponse(header = NULL, mutationResult = NULL)
}
\arguments{
\item{header}{No description}

\item{mutationResult}{The result of performing the mutation (always present)}
}
\value{
BlindWriteResponse object
}
\description{
BlindWriteResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

