% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{CommitResponse}
\alias{CommitResponse}
\title{CommitResponse Object}
\usage{
CommitResponse(header = NULL, mutationResult = NULL)
}
\arguments{
\item{header}{No description}

\item{mutationResult}{The result of performing the mutation (if any)}
}
\value{
CommitResponse object
}
\description{
CommitResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

