% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{CompositeFilter}
\alias{CompositeFilter}
\title{CompositeFilter Object}
\usage{
CompositeFilter(filters = NULL, operator = NULL)
}
\arguments{
\item{filters}{The list of filters to combine}

\item{operator}{The operator for combining multiple filters}
}
\value{
CompositeFilter object
}
\description{
CompositeFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A filter that merges the multiple other filters using the given operation.
}
