% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_functions.R
\docType{package}
\name{datastore_googleAuthR}
\alias{datastore_googleAuthR}
\alias{datastore_googleAuthR-package}
\title{Google Cloud Datastore API
Stores and queries data in Google Cloud Datastore.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:05:20
filename: /Users/mark/dev/R/autoGoogleAPI/googledatastorev1beta1.auto/R/datastore_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
\item https://www.googleapis.com/auth/userinfo.email
}
}

