% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Entity}
\alias{Entity}
\title{Entity Object}
\usage{
Entity(Entity.properties = NULL, key = NULL, properties = NULL)
}
\arguments{
\item{Entity.properties}{The \link{Entity.properties} object or list of objects}

\item{key}{The entity's key}

\item{properties}{The entity's properties}
}
\value{
Entity object
}
\description{
Entity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entity.
}
\seealso{
Other Entity functions: \code{\link{Entity.properties}}
}
