% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Filter}
\alias{Filter}
\title{Filter Object}
\usage{
Filter(compositeFilter = NULL, propertyFilter = NULL)
}
\arguments{
\item{compositeFilter}{A composite filter}

\item{propertyFilter}{A filter on a property}
}
\value{
Filter object
}
\description{
Filter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A holder for any type of filter. Exactly one field should be specified.
}

