% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{GqlQuery}
\alias{GqlQuery}
\title{GqlQuery Object}
\usage{
GqlQuery(allowLiteral = NULL, nameArgs = NULL, numberArgs = NULL,
  queryString = NULL)
}
\arguments{
\item{allowLiteral}{When false, the query string must not contain a literal}

\item{nameArgs}{A named argument must set field GqlQueryArg}

\item{numberArgs}{Numbered binding site @1 references the first numbered argument, effectively using 1-based indexing, rather than the usual 0}

\item{queryString}{The query string}
}
\value{
GqlQuery object
}
\description{
GqlQuery Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A GQL query.
}
