% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Key}
\alias{Key}
\title{Key Object}
\usage{
Key(partitionId = NULL, path = NULL)
}
\arguments{
\item{partitionId}{Entities are partitioned into subsets, currently identified by a dataset (usually implicitly specified by the project) and namespace ID}

\item{path}{The entity path}
}
\value{
Key object
}
\description{
Key Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A unique identifier for an entity.
}
