% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{KeyPathElement}
\alias{KeyPathElement}
\title{KeyPathElement Object}
\usage{
KeyPathElement(id = NULL, name = NULL)
}
\arguments{
\item{id}{The ID of the entity}

\item{name}{The name of the entity}
}
\value{
KeyPathElement object
}
\description{
KeyPathElement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A (kind, ID/name) pair used to construct a key path.At most one of name or ID may be set. If either is set, the element is complete. If neither is set, the element is incomplete.
}

