% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{LookupRequest}
\alias{LookupRequest}
\title{LookupRequest Object}
\usage{
LookupRequest(keys = NULL, readOptions = NULL)
}
\arguments{
\item{keys}{Keys of entities to look up from the datastore}

\item{readOptions}{Options for this lookup request}
}
\value{
LookupRequest object
}
\description{
LookupRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other LookupRequest functions: \code{\link{datasets.lookup}}
}

