% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{MutationResult}
\alias{MutationResult}
\title{MutationResult Object}
\usage{
MutationResult(indexUpdates = NULL, insertAutoIdKeys = NULL)
}
\arguments{
\item{indexUpdates}{Number of index writes}

\item{insertAutoIdKeys}{Keys for insertAutoId entities}
}
\value{
MutationResult object
}
\description{
MutationResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

