% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Property}
\alias{Property}
\title{Property Object}
\usage{
Property(multi = NULL, values = NULL)
}
\arguments{
\item{multi}{If this property contains a list of values}

\item{values}{The value(s) of the property}
}
\value{
Property object
}
\description{
Property Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entity property.
}

