% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Query}
\alias{Query}
\title{Query Object}
\usage{
Query(endCursor = NULL, filter = NULL, groupBy = NULL, kinds = NULL,
  limit = NULL, offset = NULL, order = NULL, projection = NULL,
  startCursor = NULL)
}
\arguments{
\item{endCursor}{An ending point for the query results}

\item{filter}{The filter to apply (optional)}

\item{groupBy}{The properties to group by (if empty, no grouping is applied to the result set)}

\item{kinds}{The kinds to query (if empty, returns entities from all kinds)}

\item{limit}{The maximum number of results to return}

\item{offset}{The number of results to skip}

\item{order}{The order to apply to the query results (if empty, order is unspecified)}

\item{projection}{The projection to return}

\item{startCursor}{A starting point for the query results}
}
\value{
Query object
}
\description{
Query Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A query.
}

