% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{QueryResultBatch}
\alias{QueryResultBatch}
\title{QueryResultBatch Object}
\usage{
QueryResultBatch(endCursor = NULL, entityResultType = NULL,
  entityResults = NULL, moreResults = NULL, skippedResults = NULL)
}
\arguments{
\item{endCursor}{A cursor that points to the position after the last result in the batch}

\item{entityResultType}{The result type for every entity in entityResults}

\item{entityResults}{The results for this batch}

\item{moreResults}{The state of the query after the current batch}

\item{skippedResults}{The number of results skipped because of Query}
}
\value{
QueryResultBatch object
}
\description{
QueryResultBatch Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch of results produced by a query.
}
