% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{ReadOptions}
\alias{ReadOptions}
\title{ReadOptions Object}
\usage{
ReadOptions(readConsistency = NULL, transaction = NULL)
}
\arguments{
\item{readConsistency}{The read consistency to use}

\item{transaction}{The transaction to use}
}
\value{
ReadOptions object
}
\description{
ReadOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

