% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{RollbackRequest}
\alias{RollbackRequest}
\title{RollbackRequest Object}
\usage{
RollbackRequest(transaction = NULL)
}
\arguments{
\item{transaction}{The transaction identifier, returned by a call to beginTransaction}
}
\value{
RollbackRequest object
}
\description{
RollbackRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RollbackRequest functions: \code{\link{datasets.rollback}}
}

