% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{RunQueryResponse}
\alias{RunQueryResponse}
\title{RunQueryResponse Object}
\usage{
RunQueryResponse(batch = NULL, header = NULL)
}
\arguments{
\item{batch}{A batch of query results (always present)}

\item{header}{No description}
}
\value{
RunQueryResponse object
}
\description{
RunQueryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

